pyproject.toml
[project]
name = "mygarden"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "pyjwt>=2.10.1",
    "pymongo>=4.15.0",
]
